requirements.txt
aiofiles
dill>=0.3.7
dnspython>=2.4.0
evidently>=0.4.0
Flask
httptools
imblearn
jinja2>=3.1.3
matplotlib
mypy-boto3-s3
Werkzeug>=3.0.0
pip-chill
pymongo>=4.6.0
python-dotenv
python-multipart
seaborn
types-s3transfer
uvicorn>=0.23.0
watchfiles
websockets>=11.0
wincertstore
xgboost>=2.0.0
neuro-mf
boto3>=1.28.0
fastapi>=0.100.0
nltk
numpy
pandas>=2.1.0
setuptools
ipython
-e .